[Breaking] Add secret config property
