+++ src/docfx/restore/PackageResolver.cs
@@ -150,7 +150,7 @@ namespace Microsoft.Docs.Build
                     {
                         Log.Write($"{committish} branch doesn't exist on repository {url}, fallback to {branch} branch");
                     }
-                    GitUtility.Fetch(_config, cwd, url, $"+{branch}:{branch}", $"{fetchOption} {depthOneOption}");
+                    GitUtility.Fetch(_config.Secrets, cwd, url, $"+{branch}:{branch}", $"{fetchOption} {depthOneOption}");
                     succeeded = true;
                     branchUsed = branch;
                     break;
@@ -165,7 +165,7 @@ namespace Microsoft.Docs.Build
                 try
                 {
                     // Fallback to fetch all branches if the input committish is not supported by fetch
-                    GitUtility.Fetch(_config, cwd, url, "+refs/heads/*:refs/heads/*", $"{fetchOption} --depth 99999999");
+                    GitUtility.Fetch(_config.Secrets, cwd, url, "+refs/heads/*:refs/heads/*", $"{fetchOption} --depth 99999999");
                 }
                 catch (InvalidOperationException ex)
                 {
